Moved the Types.tsx fil to a common folder
Refactored imports
