+++ ClientApp/src/components/admin/overview/Recipient.tsx
@@ -3,7 +3,7 @@ import { Search } from "@material-ui/icons";
 import React, { useState } from "react";
 import useStyles from "./Styles";
 import Datatable from "../common/RecipientTable";
-import { RecipientType} from "./Types";
+import { RecipientType} from "../../../common/components/Types";
 import * as Types from "../suggestedConnections/Types";
 
 type Props = {
